storage/requirements.txt
connexion[flask,uvicorn,swagger-ui]
sqlalchemy
mysqlclient
pykafka
PyYAML
setuptools
